Temporarily disabling display of CI tests
Some of the secret teams that the Pester tests depend on
appear to no longer be valid (they were set up by the
previous module administrator).

Until this is resolved, removing the CI status from the
README in order to avoid potential end-user confusion on
the current status/quality of the module.