Fix handling of -WhatIf in core functions (#213)
WhatIf support has now been simiplified in all API's that eventually
call into `Invoke-GHRestMethod`.  There's now a single check at the
top of that function which checks if it should continue or not.  If it
shouldn't, it early returns in order to avoid any code that might access
uninitialized content from outside of the ShouldProcess blocks later on.

Resolves #197