Fix default logPath when no user profile is available (#283)
In Azure Function Apps, the special `MyDocuments` folder resolves to an empty string, which means that the default `logPath` gets stored as an empty string.

https://github.com/microsoft/PowerShellForGitHub/blob/3094c9789caf00382134a0703344f655b13a1e33/GitHubConfiguration.ps1#L650-L657

https://github.com/microsoft/PowerShellForGitHub/blob/3094c9789caf00382134a0703344f655b13a1e33/GitHubConfiguration.ps1#L671

This causes problems when `Write-Log` is called by any other function, because we directly assign the current configuration value for `LogPath` as the parameter's default, which throws an exception when an empty string is assigned.

https://github.com/microsoft/PowerShellForGitHub/blob/3094c9789caf00382134a0703344f655b13a1e33/Helpers.ps1#L142

In this scenario, we'll fall back to using the `LocalApplicationDataFolder` folder and writing a verbose message to the user letting them know this has happened.  I've opted for a verbose message as opposed to a warning to avoid a constant annoying warning for users.  The alternate path is documented in the updated USAGE.md, and it's also retrievable by calling `Get-GitHubConfiguration -Name LogPath`.

Documentation has also been updated to reflect this behavior change.

Fixes #282